requirements.txt
streamlit
langchain
pypdf
openai
python-dotenv
tik-token